\name{GHSCM-package}
\alias{GHSCM-package}
\alias{GHSCM}
\docType{package}
\title{
\packageTitle{GHSCM}
}
\description{
\packageDescription{GHSCM}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{GHSCM}

Maintainer: \packageMaintainer{GHSCM}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = GHSCM)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{GHSCM}}
%\section{Documentation Index}{\packageIndices{GHSCM}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
